+++ example/lib/video_list.dart
@@ -38,7 +38,7 @@ class _VideoListState extends State<VideoList> {
       body: ListView.separated(
         itemBuilder: (context, index) {
           return YoutubePlayer(
-            key: ValueKey(_controllers[index].hashCode),
+            key: ObjectKey(_controllers[index]),
             controller: _controllers[index],
             actionsPadding: EdgeInsets.only(left: 16.0),
             bottomActions: [