+++ lib/src/widgets/duration_widgets.dart
@@ -107,7 +107,7 @@ class _RemainingDurationState extends State<RemainingDuration> {
   Widget build(BuildContext context) {
     return Text(
       "- ${durationFormatter(
-        (_controller.value.duration?.inMilliseconds ?? 0) -
+        (_controller.metadata.duration?.inMilliseconds ?? 0) -
             (_controller.value.position?.inMilliseconds ?? 0),
       )}",
       style: TextStyle(